Basic program that outputs the total dozen of eggs a user has given the total amount of eggs

+++ eggs.c
@@ -7,7 +7,7 @@ int main()
     int eggs;
     scanf("%i", &eggs);
 
-    double dozen = eggs / 12;
+    double dozen = (double) eggs / 12;
 
     printf("You have %f dozen eggs.\n", dozen);
